# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DiscordUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  avatarUrl: string;
  /** @nullable */
  email?: string | null;
}

export interface BotStats {
  guildCount: number;
  userCount: number;
  commandCount: number;
  /** Uptime in seconds */
  uptime: number;
  /** Bot latency in ms */
  ping: number;
  botName: string;
  /** @nullable */
  botAvatarUrl?: string | null;
  botId: string;
}

export interface BotGuild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  memberCount: number;
  ownerId: string;
}

export interface UserGuild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  botInstalled: boolean;
  permissions: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}