# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BotStats {
  guildCount: number;
  userCount: number;
  commandCount: number;
  /** Uptime in seconds */
  uptime: number;
  /** Bot latency in ms */
  ping: number;
  botName: string;
  /** @nullable */
  botAvatarUrl?: string | null;
  botId: string;
}