# lib/api-zod/src/generated/types/discordUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordUser {
  id: string;
  username: string;
  discriminator: string;
  /** @nullable */
  avatar?: string | null;
  avatarUrl: string;
  /** @nullable */
  email?: string | null;
}