# lib/api-zod/src/generated/types/userGuild.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Discord Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UserGuild {
  id: string;
  name: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  iconUrl?: string | null;
  botInstalled: boolean;
  permissions: string;
}